// backend/src/domain/courses/enums/CourseErrorType.ts
export enum CourseErrorType {
    InvalidCourseId = "Invalid course ID",
    InvalidCourseTitle = "Invalid course title",
    InvalidSpecialization = "Invalid specialization",
    InvalidFaculty = "Invalid faculty",
    InvalidCredits = "Invalid credits",
    InvalidEnrollmentId = "Invalid enrollment ID",
    InvalidStudentId = "Invalid student ID",
    InvalidReason = "Invalid reason",
    CourseNotFound = "Course not found",
    EnrollmentNotFound = "Enrollment not found",
    EnrollmentNotPending = "Enrollment is not in pending status",
    CourseFull = "Course has reached maximum enrollment",
  }